src/update_final_table/requirements.txt
pandas
psycopg2-binary
clickhouse-driver
SQLAlchemy
clickhouse-sqlalchemy
